requirements.txt
uvicorn~=0.24.0
loguru~=0.7.2
fastapi~=0.104.1
SQLAlchemy~=2.0.23
pydantic~=2.4.2
PyJWT~=2.8.0
httpx~=0.24.1
schedule~=1.2.1
pillow~=10.0.0